tests/test_api.py
"""Tests for FastAPI endpoints."""

import pytest
from fastapi.testclient import TestClient
import numpy as np
import cv2
import io


class TestHealthEndpoint:
    """Tests for health check endpoint."""
    
    def test_health_check_structure(self):
        # Import here to avoid loading model during test discovery
        from api.main import app
        
        # Note: In real tests, you'd use a test client with mocked model
        # For now, just verify the endpoint exists
        assert app is not None


class TestAnalysisEndpoints:
    """Tests for analysis endpoints."""
    
    @pytest.fixture
    def sample_image(self):
        """Create a sample test image."""
        image = np.random.randint(0, 256, (224, 224, 3), dtype=np.uint8)
        _, buffer = cv2.imencode('.png', image)
        return io.BytesIO(buffer.tobytes())
    
    def test_sample_image_creation(self, sample_image):
        """Test that sample image is created correctly."""
        assert sample_image is not None
        content = sample_image.read()
        assert len(content) > 0
