pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uterus-scope-ai"
version = "0.1.0"
description = "Autonomous Ultrasound Interpretation Agent for endometrial analysis"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "UterusScope-AI Team" }
]
keywords = [
    "medical-ai",
    "ultrasound",
    "vision-transformer",
    "endometrial-analysis",
    "monai",
    "clinical-decision-support"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core ML
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "monai>=1.3.0",
    "timm>=0.9.0",
    
    # Image processing
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    
    # API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    
    # Export & Inference
    "onnx>=1.15.0",
    "onnxruntime>=1.16.0",
    
    # Reporting
    "jinja2>=3.1.0",
    "matplotlib>=3.8.0",
    "weasyprint>=60.0",
    
    # Utilities
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
uterus-scope = "uterus_scope.cli:main"

[project.urls]
Homepage = "https://github.com/uterus-scope-ai/uterus-scope-ai"
Documentation = "https://uterus-scope-ai.readthedocs.io"
Repository = "https://github.com/uterus-scope-ai/uterus-scope-ai"

[tool.hatch.build.targets.wheel]
packages = ["src/uterus_scope"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["uterus_scope"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
